requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==25.2
Pillow==12.2.0
openai==2.30.0
python-dotenv==1.0.0
celery==5.6.3
redis==5.0.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.15.1
boto3==1.42.83
django-storages==1.14.6
requests==2.33.1
setuptools
